requirements.txt
# Core dependencies for RuralAssist
faiss-cpu>=1.7.0
pdfminer.six>=20221105
beautifulsoup4>=4.11.0
sentence-transformers>=2.2.0
numpy>=1.23.0
scipy>=1.9.0
lxml>=4.9.0
requests>=2.28.0
pandas>=1.5.0

# Configuration
pyyaml>=6.0

# Environment variables
python-dotenv>=0.21.0

# Testing
pytest>=7.0
pytest-cov>=4.0
pytest-timeout>=2.1.0

# Development & Linting (optional)
black>=23.0
flake8>=6.0
mypy>=1.0